renamed `db` -> `orm`, refactoring
